Update README.md
This Instagram Clone is a front-end project created to practice web development skills and understand the core concepts of HTML and CSS. The goal was to replicate the look and feel of Instagram while keeping the design simple and functional. The project focuses on building a responsive layout using Flexbox and Grid, emphasizing the importance of maintaining consistency in design and usability. It includes a feed, user profile section, and interactive elements, such as like buttons and comment sections. Through this project, I learned how to implement real-world design challenges, improve CSS styling techniques, and enhance the overall user experience. This project is a step towards gaining a deeper understanding of web development and improving my front-end skills.